Create a conversion function for textures
* Break out methods so they're more modular
* Implement the `From` trait for SerializedTextured -> Textured
